refactor(file): extract AI-policy codecs to submodule; widen RQL AI-policy coverage
- physical_metadata.rs was 2121 lines on main (already over the 2000-line
  TigerStyle file-contract budget; the layout_authority guard only runs on PRs,
  so direct-to-main pushes never tripped it). Move the AI-policy JSON codecs
  (verbatim) into physical_metadata/ai_policy_codec.rs → parent now 1967 lines.
- Add EMBED/VISION error-branch + alias tests in reddb-io-rql to lift crate line
  coverage off the 90% knife-edge.

Refs #1274
